Fix timer function, fix slow slider in classic reloaded, fix loop recording
git-svn-id: svn://code.dyne.org/veejay/trunk@1231 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-current/veejay-client/src/utils.h
@@ -20,6 +20,6 @@
 #ifndef GVRUTILS
 #define GVRUTILS
 int 	status_to_arr( char *status, int *array );
-char   *format_time( int pos , float fps);
+char   *format_time( int pos , double fps);
 #endif
